Revert "Add Vitest for unit tests" (#35)
This reverts commit 52aa813616d3b69dfb3abf466208730392c81ec6.
