+++ .gitignore
@@ -8,6 +8,3 @@ scripts/*
 *.js.map
 *.tsbuildinfo
 **/dist/**/*.d.ts
-
-# Vitest
-coverage
