+++ .vscode/settings.json
@@ -2,5 +2,6 @@
 	"editor.insertSpaces": false,
 	"typescript.tsc.autoDetect": "off",
 	"typescript.preferences.quoteStyle": "single",
-	"git.ignoreLimitWarning": true
+	"git.ignoreLimitWarning": true,
+	"editor.formatOnSave": false
 }
