--- .vscode/snippets.code-snippets
@@ -1,40 +1,0 @@
-{
-	// Place your ahk2 workspace snippets here. Each snippet is defined under a snippet name and has a scope, prefix, body and
-	// description. Add comma separated ids of the languages where the snippet is applicable in the scope field. If scope
-	// is left empty or omitted, the snippet gets applied to all languages. The prefix is what is
-	// used to trigger the snippet and the body will be expanded and inserted. Possible variables are:
-	// $1, $2 for tab stops, $0 for the final cursor position, and ${1:label}, ${2:another} for placeholders.
-	// Placeholders with the same ids are connected.
-	// Example:
-	// "Print to console": {
-	// 	"scope": "javascript,typescript",
-	// 	"prefix": "log",
-	// 	"body": [
-	// 		"console.log('$1');",
-	// 		"$2"
-	// 	],
-	// 	"description": "Log output to console"
-	// }
-
-	"Add unit test": {
-		"scope": "typescript",
-		"prefix": "describe",
-		"body": [
-			"describe('$1', () => {",
-			"\ttest.concurrent.each<",
-			"\t[",
-			"\t\tname: string,",
-			"\t\t\targs: Parameters<typeof $1>,",
-			"\t\t\texpected: ReturnType<typeof $1>,",
-			"\t\t]",
-			"\t>([",
-			"\t['$2', [$3], $4],",
-			"\t])('%s', (_name, args, expected) => {",
-			"\t\tconst result = $1(...args);",
-			"\t\texpect(result).toBe(expected);",
-			"\t});",
-			"});",
-		],
-		"description": "Add Vitest test.concurrent.each in a describe block",
-	},
-}
