+++ package.json
@@ -49,17 +49,14 @@
 		"lint:fix": "npm run format:fix && npm run eslint:fix && npm run sort-package-json:fix",
 		"sort-package-json": "sort-package-json --check",
 		"sort-package-json:fix": "sort-package-json",
-		"test": "npm run test:e2e",
-		"test//": "echo Run e2e tests (open VS Code), not grammar tests",
-		"test:ci": "npm run test:unit",
-		"pretest:e2e": "npm run compile-ts",
-		"test:e2e": "cd client && vscode-test",
-		"test:e2e//": "echo todo failing for now, will be adding pure unit tests soon",
-		"test:unit": "vitest --run",
+		"test": "npm run test-unit",
+		"pretest-unit": "npm run compile-ts",
+		"test-unit": "cd client && vscode-test",
+		"test-unit//": "echo todo failing for now, will be adding pure unit tests soon",
+		"test//": "Run unit tests, not grammar tests",
 		"validate": "npm run lint && npm run test",
-		"validate:ci": "npm run lint && npm run test:ci",
-		"validate:ci:fix": "npm run lint:fix && npm run test:ci",
-		"validate:ci:fix//": "echo Fix validation issues first, then run remaining checks until failure",
+		"validate:ci": "npm run lint",
+		"validate:ci:fix": "npm run lint:fix",
 		"watch": "tsc -b -w"
 	},
 	"prettier": {
@@ -90,7 +87,6 @@
 		"@types/node": "^20.16.0",
 		"@types/sinon": "^17.0.3",
 		"@types/vscode": "^1.82.0",
-		"@vitest/coverage-v8": "^2.1.1",
 		"@vscode/test-cli": "^0.0.10",
 		"@vscode/test-electron": "^2.4.1",
 		"@vscode/test-web": "^0.0.56",
@@ -105,7 +101,6 @@
 		"ts-loader": "^9.4.0",
 		"typescript": "^5.3.2",
 		"typescript-eslint": "^7.12.0",
-		"vitest": "^2.1.1",
 		"vscode-tmgrammar-test": "^0.1.3",
 		"webpack": "^5.74.0",
 		"webpack-cli": "^4.10.0"
